cmd/charm/charmcmd: simplify plugin processing
With the help of a small addition to the juju/cmd API
we can avoid the need to register plugin commands in
advance, thus speeding up the charm command in the
common case. As a bonus this means we don't need
any of the custom (mostly duplicated) code to layer
on top of cmd.SuperCommand, so we can delete all that.