use bakery.v2
We change the client code to use bakery.v2 (which is backwardly
compatible with the old bakery) while leaving the charm store server code
itself at v1 (it's too hard to change for now).

This will allow us to use the new agent code in the charm command
(upcoming PR).

Some of the dependencies still use bakery.v1 (notably terms-client)
we'll leave those for now.