cmd/ingest: core ingestion logic
We implement the ingest function to copy the actual required
content. This is still all only tested using a fake charmstore
implementation, so might all fall down when used for
real, but the basic logic should be reasonably sound
and the interface should be trivial to implement in terms
of the public charmrepo/csclient API.

Main missing feature so far: resources.